refactor: delete  semester form values mapper
