+++ frontend/src/components/Semester/SemesterSidebar.js
@@ -3,7 +3,6 @@ import SearchPanel from '../../share/SearchPanel/SearchPanel';
 import SemesterForm from '../../containers/SemesterPage/SemesterForm';
 import i18n from '../../i18n';
 import { COMMON_SEMESTER_IS_NOT_UNIQUE } from '../../constants/translationLabels/common';
-import { semesterFormValueMapper } from '../../helper/semesterFormValueMapper';
 import { checkUniqSemester } from '../../validation/storeValidation';
 import { checkSemesterYears } from '../../utils/formUtils';
 
@@ -18,26 +17,46 @@ const SemesterSidebar = (props) => {
         handleSemester,
         semester,
         options,
+        classScheduler,
     } = props;
 
     const [selectedGroups, setSelectedGroups] = useState([]);
 
     const submitSemesterForm = (values) => {
+        const semesterDays = [];
+        const semesterClasses = [];
         const semesterGroups = selectedGroups.map((group) => {
             return { id: group.id, title: group.label };
         });
-        const formValues = { ...values, semester_groups: semesterGroups };
-        const semester = semesterFormValueMapper(formValues);
-        if (!checkUniqSemester(semester)) {
+
+        Object.keys(values).forEach((prop) => {
+            if (prop.indexOf('semester_days_markup_') >= 0 && values[prop] === true) {
+                semesterDays.push(prop.substring(21));
+            }
+            if (prop.indexOf('semester_classes_markup_') >= 0 && values[prop] === true) {
+                semesterClasses.push(
+                    classScheduler.find((schedule) => schedule.id === +prop.substring(24)),
+                );
+            }
+        });
+        const semesterItem = {
+            ...values,
+            semester_groups: semesterGroups,
+            semester_classes: semesterClasses,
+            semester_days: semesterDays,
+        };
+        if (!checkUniqSemester(semesterItem)) {
             const message = i18n.t(COMMON_SEMESTER_IS_NOT_UNIQUE);
             setOpenErrorSnackbar(message);
             setError(true);
             return;
         }
-        if (!checkSemesterYears(semester.endDay, semester.startDay, semester.year)) return;
-        handleSemester(semester);
+        if (!checkSemesterYears(semesterItem.endDay, semesterItem.startDay, semesterItem.year))
+            return;
+        handleSemester(semesterItem);
         setSelectedGroups([]);
     };
+
     return (
         <aside className="search-list__panel">
             <SearchPanel
